LCEL/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["syedmuneeb321 <130203424+syedmuneeb321@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.4"}
langserve = {extras = ["all"], version = "^0.3.0"}
langchain-groq = "^0.2.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
